+++ src/main/java/vectorwing/farmersdelight/items/HorseFeedItem.java
@@ -25,14 +25,13 @@ import net.minecraftforge.fml.common.Mod;
 import vectorwing.farmersdelight.FarmersDelight;
 import vectorwing.farmersdelight.registry.ModItems;
 import vectorwing.farmersdelight.registry.ModParticleTypes;
+import vectorwing.farmersdelight.setup.Configuration;
 import vectorwing.farmersdelight.utils.MathUtils;
 import vectorwing.farmersdelight.utils.TextUtils;
 import vectorwing.farmersdelight.utils.tags.ModTags;
 
 import javax.annotation.Nullable;
 import java.util.List;
-
-import net.minecraft.item.Item.Properties;
 
 public class HorseFeedItem extends Item
 {
@@ -85,6 +84,10 @@ public class HorseFeedItem extends Item
 
 	@Override
 	public void appendHoverText(ItemStack stack, @Nullable World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
+		if (!Configuration.FOOD_EFFECT_TOOLTIP.get()) {
+			return;
+		}
+
 		IFormattableTextComponent textWhenFeeding = TextUtils.getTranslation("tooltip.horse_feed.when_feeding");
 		tooltip.add(textWhenFeeding.withStyle(TextFormatting.GRAY));