+++ src/main/java/vectorwing/farmersdelight/setup/Configuration.java
@@ -54,8 +54,8 @@ public class Configuration
 	// CLIENT
 	public static final String CATEGORY_CLIENT = "client";
 
-	public static ForgeConfigSpec.BooleanValue NOURISHMENT_HUNGER_OVERLAY;
-	public static ForgeConfigSpec.BooleanValue MEAL_EFFECT_TOOLTIP;
+	public static ForgeConfigSpec.BooleanValue NOURISHED_HUNGER_OVERLAY;
+	public static ForgeConfigSpec.BooleanValue FOOD_EFFECT_TOOLTIP;
 
 	static {
 		ForgeConfigSpec.Builder COMMON_BUILDER = new ForgeConfigSpec.Builder();
@@ -157,10 +157,10 @@ public class Configuration
 		ForgeConfigSpec.Builder CLIENT_BUILDER = new ForgeConfigSpec.Builder();
 
 		CLIENT_BUILDER.comment("Client settings").push(CATEGORY_CLIENT);
-		NOURISHMENT_HUNGER_OVERLAY = CLIENT_BUILDER.comment("Should the hunger bar have a gilded overlay when the player has the Nourishment effect?")
+		NOURISHED_HUNGER_OVERLAY = CLIENT_BUILDER.comment("Should the hunger bar have a gilded overlay when the player has the Nourishment effect?")
 				.define("nourishmentHungerOverlay", true);
-		MEAL_EFFECT_TOOLTIP = CLIENT_BUILDER.comment("Should a meal's tooltip display which effect it provides?")
-				.define("mealEffectTooltip", true);
+		FOOD_EFFECT_TOOLTIP = CLIENT_BUILDER.comment("Should food tooltips display which effects they provide?")
+				.define("foodEffectTooltip", true);
 		CLIENT_BUILDER.pop();
 
 		CLIENT_CONFIG = CLIENT_BUILDER.build();